+++ tests/bluetooth/tester/CMakeLists.txt
@@ -12,14 +12,14 @@ zephyr_library_include_directories(${ZEPHYR_BASE}/subsys/bluetooth/mesh)
 zephyr_library_include_directories(${ZEPHYR_BASE}/subsys/bluetooth/host)
 target_sources(app PRIVATE
     src/main.c
-    src/bttester.c
-    src/gap.c
-    src/gatt.c
+    src/btp.c
+    src/btp_gap.c
+    src/btp_gatt.c
     )
 if(CONFIG_BT_L2CAP_DYNAMIC_CHANNEL)
-target_sources(app PRIVATE src/l2cap.c)
+target_sources(app PRIVATE src/btp_l2cap.c)
 endif()
 
 if(CONFIG_BT_MESH)
-target_sources(app PRIVATE src/mesh.c)
+target_sources(app PRIVATE src/btp_mesh.c)
 endif()
