+++ tests/bluetooth/tester/src/btp_gatt.c
@@ -27,7 +27,7 @@
 #define LOG_MODULE_NAME bttester_gatt
 LOG_MODULE_REGISTER(LOG_MODULE_NAME);
 
-#include "bttester.h"
+#include "btp/btp.h"
 
 #define CONTROLLER_INDEX 0
 #define MAX_BUFFER_SIZE 2048
@@ -38,14 +38,14 @@ LOG_MODULE_REGISTER(LOG_MODULE_NAME);
 #define UNUSED_SUBSCRIBE_CCC_HANDLE 0x0000
 
 /* This masks Permission bits from GATT API */
-#define GATT_PERM_MASK			(BT_GATT_PERM_READ | \
+#define GATT_PERM_MASK		(BT_GATT_PERM_READ | \
 					 BT_GATT_PERM_READ_AUTHEN | \
 					 BT_GATT_PERM_READ_ENCRYPT | \
 					 BT_GATT_PERM_WRITE | \
 					 BT_GATT_PERM_WRITE_AUTHEN | \
 					 BT_GATT_PERM_WRITE_ENCRYPT | \
 					 BT_GATT_PERM_PREPARE_WRITE)
-#define GATT_PERM_ENC_READ_MASK		(BT_GATT_PERM_READ_ENCRYPT | \
+#define GATT_PERM_ENC_READ_MASK	(BT_GATT_PERM_READ_ENCRYPT | \
 					 BT_GATT_PERM_READ_AUTHEN)
 #define GATT_PERM_ENC_WRITE_MASK	(BT_GATT_PERM_WRITE_ENCRYPT | \
 					 BT_GATT_PERM_WRITE_AUTHEN)
